rename to pyo3-stubgen
